Revert "passed in rule properly"
This reverts commit deb1910b00d3fa4733be95f6e636091a103f06f5.
