+++ src/PresentationalComponents/Inventory/Hooks/useBulkSelect/helpers.js
@@ -1,6 +1,8 @@
-export const compileTitle = (itemsTotal) => {
-  if (itemsTotal === 0) {
-    return null;
+export const compileTitle = (itemsTotal, titleOption) => {
+  if (typeof titleOption === 'string') {
+    return titleOption;
+  } else if (typeof titleOption === 'function') {
+    return titleOption(itemsTotal);
   } else {
     return `${itemsTotal} selected`;
   }
